backend/requirements.txt
Flask
flask-cors
PyJWT
mysql-connector-python
python-dotenv
Pillow
werkzeug
requests
langchain 
huggingface_hub 
transformers 
accelerate 
bitsandbytes 
langchain-huggingface
langchain-community
pandas
scikit-learn
numpy<2
SQLAlchemy
PyMySQL
scikit-surprise
torch
sentence-transformers
langchain-groq
langgraph